refactor display quiz into folders
