+++ src/interface_adapter/quiz/display_quiz/DisplayQuizController.java
@@ -1,4 +1,4 @@
-package interface_adapter.quiz;
+package interface_adapter.quiz.display_quiz;
 
 import use_case.quiz.display_quiz.DisplayQuizInputBoundary;
